requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.89.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0
markupsafe==2.0.1
apache-airflow-providers-odbc
pyodbc
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-microsoft-mssql[odbc]
apache-airflow-providers-microsoft-azure
gitpython
selenium
clickhouse_driver
numpy
pandas
requests